Move LighthouseApiImpl into Types.Interfaces
